io_uring: fix proptest user_data type
The proptest was previously creating a pointer to the
wrong type, which is incorrect and undefined behaviour.

+++ src/io_uring/src/lib.rs
@@ -350,7 +350,7 @@ mod tests {
     use vm_memory::{Bytes, MmapRegion, VolatileMemory};
 
     fn drain_cqueue(ring: &mut IoUring) {
-        while let Some(entry) = ring.pop::<usize>().unwrap() {
+        while let Some(entry) = ring.pop::<u32>().unwrap() {
             assert!(entry.result().is_ok());
 
             // Assert that there were no partial writes.
